+++ src/main/java/cap/s42academy/service/impl/CustomerServiceImpl.java
@@ -1,13 +1,17 @@
 package cap.s42academy.service.impl;
 
+import cap.s42academy.exceptions.BusinessException;
 import cap.s42academy.exceptions.CustomerNotFoundException;
+import cap.s42academy.model.CarRental;
 import cap.s42academy.model.Customer;
+import cap.s42academy.repository.CarRentalRepository;
 import cap.s42academy.repository.CustomerRepository;
 import cap.s42academy.service.CustomerService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.List;
 import java.util.Optional;
 
 @Service
@@ -16,6 +20,7 @@ import java.util.Optional;
 public class CustomerServiceImpl implements CustomerService {
 
     private final CustomerRepository customerRepository;
+    private final CarRentalRepository carRentalRepository;
 
     @Override
     @Transactional(readOnly = true)
@@ -44,9 +49,13 @@ public class CustomerServiceImpl implements CustomerService {
 
     @Override
     public void deleteCustomer(Long customerId) {
-        //dodac walidacje, ze jezeli klient ma rezerwacje aktualna badz w przyszlosci nie mozna go usunac
         Customer customer = getCustomer(customerId)
                 .orElseThrow(() -> new CustomerNotFoundException(customerId));
+        List<CarRental> rentListCustomer = carRentalRepository.checkBookingByCustomer(customerId);
+        if (!rentListCustomer.isEmpty()) {
+            throw new BusinessException(String.format(
+                    "Customer with id: %s has an active reservation!", customerId));
+        }
         customerRepository.delete(customer);
     }
